exam-backend/requirements.txt
asgiref==3.10.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
